Extract PriorityButtonsBar to a dedicated component
