+++ src/components/editor/PriorityButtonsBar.scss
@@ -1,0 +1,23 @@
+@import "css/common/vars/fonts";
+@import "./common.scss";
+
+.self {
+  padding: 0 $editor-horizontal-spacing 0 $editor-horizontal-spacing;
+  margin: $editor-vertical-spacing 0 $editor-vertical-spacing 0;
+
+  .priorityHeader {
+    display: inline-block;
+    font-weight: $font-semi-bold;
+    margin: 0 0 10px 0;
+  }
+
+  .headerButtons {
+    display: flex;
+    justify-content: space-between;
+    flex: 1;
+
+    :not(:first-child):not(:last-child) {
+      margin: 0 10px;
+    }
+  }
+}
