+++ src/components/editor/common.scss
@@ -1,0 +1,2 @@
+$editor-horizontal-spacing: 20px;
+$editor-vertical-spacing: 15px;
